requirements.txt
torch>=1.13.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
tqdm>=4.62.0
seaborn>=0.11.0
datasets>=2.0.0
